+++ horde/classes/base/worker.py
@@ -60,7 +60,7 @@ class WorkerTemplate(db.Model):
         "polymorphic_identity": "worker_template",
         "polymorphic_on": "worker_type",
     }    
-    suspicion_threshold = 3
+    suspicion_threshold = 5
     # Every how many seconds does this worker get a kudos reward
     uptime_reward_threshold = 600
     default_maintenance_msg = "This worker has been put into maintenance mode by its owner"
@@ -158,6 +158,7 @@ class WorkerTemplate(db.Model):
         if int(reason) in self.suspicions and reason not in [Suspicions.UNREASONABLY_FAST,Suspicions.TOO_MANY_JOBS_ABORTED]:
             return
         new_suspicion = WorkerSuspicions(worker_id=self.id, suspicion_id=int(reason))
+        db.session.add(new_suspicion)
         self.user.report_suspicion(amount, reason, formats)
         if reason:
             reason_log = SUSPICION_LOGS[reason].format(*formats)
@@ -172,7 +173,7 @@ class WorkerTemplate(db.Model):
         db.session.commit()   
 
     def get_suspicion(self):
-        return(len(self.suspicions))
+        return len(self.suspicions)
 
     def is_suspicious(self):
         # Trusted users are never suspicious
@@ -273,11 +274,16 @@ class WorkerTemplate(db.Model):
         # We reurn the converted amount as well in case we need it
         return(converted)
 
+    def get_bridge_kudos_multiplier(self):
+        '''To override in case we want to adjust the worker reward based on their bridge version'''
+        return 1
+
     @logger.catch(reraise=True)
     def record_contribution(self, raw_things, kudos, things_per_sec):
         '''We record the servers newest contribution
         We do not need to know what type the contribution is, to avoid unnecessarily extending this method
         '''
+        kudos = kudos * self.get_bridge_kudos_multiplier()
         self.user.record_contributions(raw_things = raw_things, kudos = kudos, contrib_type = self.wtype)
         self.modify_kudos(kudos,'generated')
         converted_amount = self.convert_contribution(raw_things)
@@ -586,4 +592,3 @@ class Worker(WorkerTemplate):
         for model in self.models:
             db.session.delete(model)
         super().delete()
-