+++ horde/consts.py
@@ -1,4 +1,4 @@
-HORDE_VERSION = "4.13.0"
+HORDE_VERSION = "4.14.0"
 
 WHITELISTED_SERVICE_IPS = {
     "212.227.227.178", # Turing Bot
@@ -6,14 +6,14 @@ WHITELISTED_SERVICE_IPS = {
 
 # And their extra kudos adjustments based on how expensive to process they are and/or how much extra horde resources they consume
 KNOWN_POST_PROCESSORS = {
-    "GFPGAN": 1.0, 
-    "RealESRGAN_x4plus": 1.3, 
-    "RealESRGAN_x2plus": 1.2, 
-    "RealESRGAN_x4plus_anime_6B": 1.3,
-    "NMKD_Siax": 1.1,
-    "4x_AnimeSharp": 1.1, 
-    "CodeFormers": 1.3, 
-    "strip_background": 1.2,
+    "GFPGAN": 1, 
+    "RealESRGAN_x4plus": 1.05, 
+    "RealESRGAN_x2plus": 1.05, 
+    "RealESRGAN_x4plus_anime_6B": 1.05,
+    "NMKD_Siax": 1.05,
+    "4x_AnimeSharp": 1.05, 
+    "CodeFormers": 1, 
+    "strip_background": 1,
 }
 
 KNOWN_UPSCALERS = [