+++ horde/database/functions.py
@@ -196,7 +196,6 @@ def find_sharedkey(shared_key):
     try:
         sharedkey_uuid = uuid.UUID(shared_key)
     except ValueError as e: 
-        logger.debug(f"Non-UUID sharedkey_id sent: '{shared_key}'.")
         return None        
     if SQLITE_MODE:
         sharedkey_uuid = str(sharedkey_uuid)